Update README.md
Made some clarifying edits to the readme. 
May need to edit the following since I was under the impression only SNPs within structural genes were considered: 
"Both coding and non-coding variants are considered in
the genotyping analysis, with the exception of the 3’ and 5’ UTRs. All
alleles of SNPs with a significant minor allele are visualized, and
compared to the alleles of the user uploaded viral genome."